feat : 프로젝트 페이지 생성 준비, project 하위 경로(폴더) 생성
