+++ src/app/project/CreateProject/CreateProject.styles.tsx
@@ -1,4 +1,4 @@
-import styled from 'styled-components';
+import styled from '@emotion/styled';
 
 export const CreateProjectContainer = styled.div`
   display: flex;
