+++ src/app/project/InvitedProject.tsx
@@ -1,0 +1,13 @@
+import { ProjectView } from './Project.styles';
+
+const InvitedProject = () => {
+  return (
+    <ProjectView>
+      <div>
+        <h2>초대받은 프로젝트</h2>
+      </div>
+    </ProjectView>
+  );
+};
+
+export default InvitedProject;
